esp32/sdkconfig: Change to use single core only.
There seem to be random crashes when using both cores at the same time.
It's hard to reproduce, but, for example, changing the order of object
files on the command line to the linker can make the difference between
crash and no crash.  And the start up code will always get at least as far
as starting the PRO CPU.

Espressif also changed the default config setting to use only a single
core.

It may be related to SPI flash and cache because the SPI flash driver
plays some tricks to get both CPUs to stop while accessing the flash.