esp/modesp.c: Add osdebug() and oslog()
Code lineage:

osdebug() is based loosely on the version in esp8266, but there didn't
seem to be an obvious way of choosing a particular UART. The basic
behavior is the same, though: provide None, and logging is disabled;
provide an integer and logging is restored to the default level.

To build on that, and because the IDF provides more functionality, a
second parameter has now been implemented which allows the active log
level to be set:

   esp.osdebug(uart[, level])

The module has a corresponding set of LOG_ values to set this
accordingly. An additional level esp.LOG_CRITICAL, which maps to
esp.LOG_ERROR has been included in order to provide better
symetry with standard Python logging frameworks.

In order to test out correct behavior, it was straightforward to add a
second function that calls the IDF to perform a logging operation:

   esp.oslog(level, tag, message)

Because our intent to write anything in Python that can be in Python,
the thought is that format and parameters will be transformed to a
simple string in Python, to them be passed here for rendering.