esp/machine_timer.c: Add esp32 hardware timer support
Code lineage:

Timer() is based loosely on the version in esp8266, although the
implementation is differs significantly because of the change in
the underlying platform.

The code provides access to the four hardware timers available
on the ESP32, where bit 1 refers to the timer group, and bit 0
refers to the timer index. It is supposed that virtual timers
could be built on top of this in Python.