tell nvs_get_str() the buffer-length; lower buffer size to 256 bytes
On nvs_get_str() the length should be initialized with the size of
the buffer.

Also lowered the size of the buffer. We should not store very large
strings in NVS, and we do not have that much space on the stack.
If we really need such large strings, we should use malloc or a
static buffer for reading the large string.