Move mount-code to modbadge; add sdcard support
The native locfd-mount was hidden in the vfs_mount code.

Mounting the fat partition and having a virtual file-system mount
to the native filesystem are two different things. Split it up.

Mounting the locfd filesystem is now done with:

    badge.mount_root()

Mounting the native filesystems into micropython is still done with:

    uos.mount(uos.VfsNative(None), '/')

Mounting the bpp moved to:

    badge.mount_bpp()

Added support for sdcard mounts:

    badge.mount_sdcard()
    badge.unmount_sdcard()

.. and fixed a printf bug in badge_eink_png.