esp32/modesp: Add esp.gpio_matrix_in and esp.gpio_matrix_out functions.
The ESP32 has a fairly complex GPIO multiplexing matrix and can do
one-to-many routing (connecting IO pins to peripherals).  When initialising
a peripheral on an IO pin the ESP IDF does not always clear existing
routing settings.  The new esp.gpio_matrix_{in,out} functions in this
commit allow the user to manually adjust the routing matrix if/when needed.

The functions also allow the user extra flexibility and control over the IO
multiplexing.